박경호 : 22-12-06 엔진 최신화

+++ GameEngineCore/GameEngineFBXAnimation.cpp
@@ -249,8 +249,8 @@ bool GameEngineFBXAnimation::AnimationLoad(std::shared_ptr <GameEngineFBXMesh> _
 				{
 					FbxExBoneFrame& Frame = CurAniData.AniFrameData[MeshIndex][boneIndex];
 					Frame.BoneMatData.resize(endTime - startTime + 1);
-					Frame.BoneIndex = boneIndex;
-					Frame.BoneParentIndex = boneIndex;
+					Frame.BoneIndex = static_cast<int>(boneIndex);
+					Frame.BoneParentIndex = static_cast<int>(boneIndex);
 
 					for (fbxsdk::FbxLongLong i = startTime; i <= endTime; ++i)
 					{
